// fe/src/lib/api/generated/model/accountsListParams.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { DesignationCode } from './designationCode';
import type { RoleCode } from './roleCode';

export type AccountsListParams = {
page?: number;
page_size?: number;
role?: RoleCode | null;
is_active?: boolean | null;
designation?: DesignationCode | null;
email_verified?: boolean | null;
search?: string | null;
};